third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/eval-scripts-setTimeout-blocked-expected.txt

CONSOLE ERROR: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'unsafe-inline'".

ALERT: PASS
